refactor: treasury, remove assethub and hydration providers and update useApi
