+++ packages/next-common/components/overview/accountInfo/accountInfoPanel.js
@@ -21,7 +21,6 @@ import dynamic from "next/dynamic";
 import { useState } from "react";
 import { OnlyChains } from "next-common/components/common/onlyChain";
 import Chains from "next-common/utils/consts/chains";
-import { AssetHubApiProvider } from "next-common/context/assetHub";
 import { RelayChainApiProvider } from "next-common/context/relayChain";
 import { CollectivesApiProvider } from "next-common/context/collectives/api";
 import useAccountUrl from "next-common/hooks/account/useAccountUrl";
@@ -286,11 +285,9 @@ export function AccountHead({ width }) {
             <TransferButton />
           </OnlyChains>
           <OnlyChains chains={relayChainTeleportEnabledChains}>
-            <AssetHubApiProvider>
-              <CollectivesApiProvider>
-                <TeleportButton />
-              </CollectivesApiProvider>
-            </AssetHubApiProvider>
+            <CollectivesApiProvider>
+              <TeleportButton />
+            </CollectivesApiProvider>
           </OnlyChains>
           <OnlyChains chains={paraChainTeleportEnabledChains}>
             <RelayChainApiProvider>
