+++ packages/next-common/components/overview/accountInfo/components/assetHubManagePrompt.js
@@ -4,7 +4,6 @@ import { CACHE_KEY } from "next-common/utils/constants";
 import { useMemo } from "react";
 import Link from "next/link";
 import useAssetsFromAssetHub from "../hook/useAssetsFromAssetHub";
-import { AssetHubApiProvider } from "next-common/context/assetHub";
 import { AssetHubMetadataProvider } from "next-common/components/overview/accountInfo/context/assetHubMetadataContext";
 import Chains from "next-common/utils/consts/chains";
 import OnlyChain from "next-common/components/common/onlyChain";
@@ -36,11 +35,9 @@ export function PromptContent() {
 export default function AssetHubManagePrompt() {
   return (
     <OnlyChain chain={Chains.polkadot}>
-      <AssetHubApiProvider>
-        <AssetHubMetadataProvider>
-          <PromptContent />
-        </AssetHubMetadataProvider>
-      </AssetHubApiProvider>
+      <AssetHubMetadataProvider>
+        <PromptContent />
+      </AssetHubMetadataProvider>
     </OnlyChain>
   );
 }
