+++ packages/next-common/components/summary/newProposalQuickStart/createUSDxTreasuryProposalPopup/index.js
@@ -5,7 +5,6 @@ import CreateProposalSubmitButton from "../common/createProposalSubmitButton";
 import { InfoMessage } from "next-common/components/setting/styled";
 import AdvanceSettings from "../common/advanceSettings";
 import BigNumber from "bignumber.js";
-import { AssetHubApiProvider } from "next-common/context/assetHub";
 import Popup from "next-common/components/popup/wrapper/Popup";
 import { useUSDxTreasuryNotePreimageTx } from "next-common/components/preImages/createPreimagePopup/templates/newUSDxTreasuryProposalPopup";
 import useUSDxBalanceField from "next-common/components/preImages/createPreimagePopup/fields/useUSDxBalanceField";
@@ -72,9 +71,7 @@ export function NewUSDxTreasuryReferendumInnerPopup() {
   const { onClose } = usePopupParams();
   return (
     <Popup title="Create USDx Treasury Proposal" onClose={onClose}>
-      <AssetHubApiProvider>
-        <PopupContent />
-      </AssetHubApiProvider>
+      <PopupContent />
     </Popup>
   );
 }
